Fix missing username for UID error.
If the UID of the user who owns the "shared" directory does not correlate to a username inside of the container, the system experiences a critical stop. This change instructs the system to check for write access in the target directory regardless of current owner, fixing this issue and potentially preventing unnecessary permission modifications.