doc->[add] code blocks under advanced section for readability.
MOTIVATION
The purpose of this update is to improve readability for future readers.
Under the *advanced* section, it was not clear what command was associated
with the `-d` option. So I included two code blocks which explicitly show
readers how to invoke the `run` command with the `-d` option, either in
MacOS/Linux or Windows.

Reported-by: Michael Sodeke
Signed-off-by: Your Name michaelsodeke@outlook.com